Revert extracting Sync interface and abstract class
