+++ src/Backup/Sync.php
@@ -1,10 +1,7 @@
 <?php
 namespace phpbu\App\Backup;
 
-use phpbu\App\Backup\Sync\Exception;
-use phpbu\App\Backup\Sync\Simulator;
 use phpbu\App\Result;
-use phpbu\App\Util;
 
 /**
  * Sync
@@ -17,21 +14,14 @@ use phpbu\App\Util;
  * @link       http://phpbu.de/
  * @since      Class available since Release 1.0.0
  */
-abstract class Sync implements Simulator
+interface Sync
 {
-    /**
-     * Unix timestamp of generating path from placeholder.
-     *
-     * @var int
-     */
-    protected $time;
-
     /**
      * Setup the Sync object with all xml options.
      *
      * @param array $options
      */
-    abstract public function setup(array $options);
+    public function setup(array $options);
 
     /**
      * Execute the Sync
@@ -40,21 +30,5 @@ abstract class Sync implements Simulator
      * @param \phpbu\App\Backup\Target $target
      * @param \phpbu\App\Result        $result
      */
-    abstract public function sync(Target $target, Result $result);
-
-    /**
-     * Make sure all mandatory keys are present in given config.
-     *
-     * @param  array $config
-     * @param  array $keys
-     * @throws Exception
-     */
-    protected function validateConfig(array $config, array $keys)
-    {
-        foreach ($keys as $option) {
-            if (!Util\Arr::isSetAndNotEmptyString($config, $option)) {
-                throw new Exception($option . ' is mandatory');
-            }
-        }
-    }
+    public function sync(Target $target, Result $result);
 }
