+++ src/Backup/Sync/Simulator.php
@@ -15,7 +15,7 @@ use phpbu\App\Result;
  * @link       http://phpbu.de/
  * @since      Class available since Release 3.0.0
  */
-interface Simulator extends \phpbu\App\Backup\Sync\Sync
+interface Simulator extends \phpbu\App\Backup\Sync
 {
     /**
      * Simulate the sync execution.
