+++ src/Factory.php
@@ -227,7 +227,7 @@ class Factory
     {
         /** @var \phpbu\App\Backup\Sync $sync */
         $sync = $this->create('sync', $alias);
-        if (!($sync instanceof \phpbu\App\Backup\Sync\Sync)) {
+        if (!($sync instanceof \phpbu\App\Backup\Sync)) {
             throw new Exception(sprintf('sync \'%s\' has to implement the \'Sync\' interface', $alias));
         }
         $sync->setup($conf);
